pyproject.toml
[project]
name = "refresh-ip-cloudflare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudflare>=3.1.0",
    "pydantic<2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[project.scripts]
hello = "hello"
